增加删除判断逻辑

+++ aiurt-module-system/src/main/java/com/aiurt/modules/position/controller/CsLineController.java
@@ -10,9 +10,15 @@ import java.net.URLDecoder;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import com.aiurt.modules.device.entity.Device;
+import com.aiurt.modules.device.service.IDeviceService;
+import com.aiurt.modules.material.entity.MaterialBaseType;
 import com.aiurt.modules.position.entity.CsLine;
+import com.aiurt.modules.position.entity.CsStation;
 import com.aiurt.modules.position.entity.CsStationPosition;
 import com.aiurt.modules.position.service.ICsLineService;
+import com.aiurt.modules.position.service.ICsStationService;
+import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
@@ -50,7 +56,10 @@ import com.aiurt.common.aspect.annotation.AutoLog;
 public class CsLineController extends BaseController<CsLine, ICsLineService> {
 	@Autowired
 	private ICsLineService csLineService;
-
+	@Autowired
+	private ICsStationService csStationService;
+	@Autowired
+	private IDeviceService deviceService;
 	/**
 	 * 分页列表查询
 	 *
@@ -125,6 +134,23 @@ public class CsLineController extends BaseController<CsLine, ICsLineService> {
 	@ApiOperation(value="cs_line-通过id删除", notes="cs_line-通过id删除")
 	@DeleteMapping(value = "/delete")
 	public Result<String> delete(@RequestParam(name="id",required=true) String id) {
+		CsLine csLine = csLineService.getById(id);
+		//判断二级是否使用
+		LambdaQueryWrapper<CsStation> wrapper =  new LambdaQueryWrapper<CsStation>();
+		wrapper.eq(CsStation::getLineCode,csLine.getLineCode());
+		wrapper.eq(CsStation::getDelFlag,0);
+		List<CsStation> list = csStationService.list(wrapper);
+		if(!list.isEmpty()){
+			return Result.error("该位置信息正在使用中，无法删除");
+		}
+		//判断设备主数据是否使用
+		LambdaQueryWrapper<Device> deviceWrapper =  new LambdaQueryWrapper<Device>();
+		deviceWrapper.eq(Device::getPositionCode,csLine.getLineCode());
+		deviceWrapper.eq(Device::getDelFlag,0);
+		List<Device> deviceList = deviceService.list(deviceWrapper);
+		if(!deviceList.isEmpty()){
+			return Result.error("该位置信息被设备主数据使用中，无法删除");
+		}
 		csLineService.removeById(id);
 		return Result.OK("删除成功!");
 	}
